+++ app/api/private-cloud/products/[licencePlate]/comments/_operations/create.ts
@@ -1,3 +1,27 @@
-export default async function createOp() {
-  return { message: 'Create comment successful' };
+import prisma from '@/core/prisma';
+
+interface CreateCommentParams {
+  text: string;
+  projectId: string;
+  userId: string;
+}
+
+export async function createOp({ text, projectId, userId }: CreateCommentParams) {
+  const comment = await prisma.privateCloudComment.create({
+    data: {
+      text,
+      user: {
+        connect: {
+          id: userId,
+        },
+      },
+      project: {
+        connect: {
+          id: projectId,
+        },
+      },
+    },
+  });
+
+  return comment;
 }