+++ queries/private-cloud-products.ts
@@ -28,6 +28,8 @@ export async function searchPrivateCloudProducts({
 }) {
   const where: Prisma.PrivateCloudProjectWhereInput = extraFilter ?? {};
 
+  if (search === '*') search = '';
+
   const orderBy =
     sort && order
       ? {
@@ -42,13 +44,18 @@ export async function searchPrivateCloudProducts({
     const productSearchcreteria: Prisma.StringFilter<'PrivateCloudProject'> = { contains: search, mode: 'insensitive' };
 
     where.OR = [
-      { projectOwnerId: { in: matchingUserIds } },
-      { primaryTechnicalLeadId: { in: matchingUserIds } },
-      { secondaryTechnicalLeadId: { in: matchingUserIds } },
       { name: productSearchcreteria },
       { description: productSearchcreteria },
       { licencePlate: productSearchcreteria },
     ];
+
+    if (matchingUserIds.length > 0) {
+      where.OR.push(
+        { projectOwnerId: { in: matchingUserIds } },
+        { primaryTechnicalLeadId: { in: matchingUserIds } },
+        { secondaryTechnicalLeadId: { in: matchingUserIds } },
+      );
+    }
   }
 
   if (ministry) {