+++ terraform/keycloak/prod/roles.tf
@@ -6,6 +6,14 @@ resource "keycloak_role" "pltsvc_admin" {
   description = "Registry Administrator"
 }
 
+resource "keycloak_role" "pltsvc_editor" {
+  realm_id  = data.keycloak_realm.pltsvc.id
+  client_id = keycloak_openid_client.pltsvc.id
+
+  name        = "editor"
+  description = "Registry Editor"
+}
+
 resource "keycloak_role" "pltsvc_reader" {
   realm_id  = data.keycloak_realm.pltsvc.id
   client_id = keycloak_openid_client.pltsvc.id
@@ -20,6 +28,14 @@ resource "keycloak_role" "pltsvc_private_admin" {
 
   name        = "private-admin"
   description = "Registry Private Cloud Administrator"
+}
+
+resource "keycloak_role" "pltsvc_private_editor" {
+  realm_id  = data.keycloak_realm.pltsvc.id
+  client_id = keycloak_openid_client.pltsvc.id
+
+  name        = "private-editor"
+  description = "Registry Private Cloud Editor"
 }
 
 resource "keycloak_role" "pltsvc_private_reader" {
@@ -38,18 +54,18 @@ resource "keycloak_role" "pltsvc_public_admin" {
   description = "Registry Public Cloud Administrator"
 }
 
+resource "keycloak_role" "pltsvc_public_editor" {
+  realm_id  = data.keycloak_realm.pltsvc.id
+  client_id = keycloak_openid_client.pltsvc.id
+
+  name        = "public-editor"
+  description = "Registry Public Cloud Editor"
+}
+
 resource "keycloak_role" "pltsvc_public_reader" {
   realm_id  = data.keycloak_realm.pltsvc.id
   client_id = keycloak_openid_client.pltsvc.id
 
   name        = "public-reader"
   description = "Registry Public Cloud Read-Only"
-}
-
-resource "keycloak_role" "pltsvc_approver" {
-  realm_id  = data.keycloak_realm.pltsvc.id
-  client_id = keycloak_openid_client.pltsvc.id
-
-  name        = "approver"
-  description = "Registry Approver"
 }